move DBConfig to conf package
